IndexedDB/tests/submissions/Opera/idbcursor_continue_index6.htm

<!DOCTYPE html>
<title>IDBCursor.continue() - index - iterate using NEXT_NO_DUPLICATE</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support.js"></script>

<script>

    var db,
      t = async_test(),
      records = [ { pKey: "primaryKey_0",   iKey: "indexKey_0" },
                  { pKey: "primaryKey_1",   iKey: "indexKey_1" },
                  { pKey: "primaryKey_1-2", iKey: "indexKey_1" },
                  { pKey: "primaryKey_2",   iKey: "indexKey_2" } ],

      actual = [ { pKey: "primaryKey_0",   iKey: "indexKey_0" },
                 { pKey: "primaryKey_1",   iKey: "indexKey_1" },
                 { pKey: "primaryKey_2",   iKey: "indexKey_2" } ];

    var open_rq = createdb(t);
    open_rq.onupgradeneeded = function(e) {
        db = e.target.result;
        var objStore = db.createObjectStore("test", { keyPath: "pKey" });

        objStore.createIndex("index", "iKey");

        for (var i = 0; i < records.length; i++)
            objStore.add(records[i]);
    };

    open_rq.onsuccess = function(e) {
        var count = 0,
          cursor_rq = db.transaction("test")
                        .objectStore("test")
                        .index("index")
                        .openCursor(undefined, IDBCursor.NEXT_NO_DUPLICATE);

        cursor_rq.onsuccess = t.step_func(function(e) {
            if (!e.target.result) {
                assert_equals(count, actual.length, 'count');
                t.done();
                return;
            }
            var cursor = e.target.result,
              record = cursor.value;

            assert_equals(record.pKey, actual[count].pKey, "pKey #" + count);
            assert_equals(record.iKey, actual[count].iKey, "iKey #" + count);

            assert_equals(cursor.key,  actual[count].iKey, "cursor.key #" + count);
            assert_equals(cursor.primaryKey, actual[count].pKey, "cursor.primaryKey #" + count);

            count++;
            cursor.continue(actual[count] ? actual[count].iKey : undefined);
        });
    };

</script>

<div id="log"></div>
